HW 4 TickerViewModel 클래스 분리
